+++ doc/changelog.rst
@@ -3,6 +3,29 @@ Changelog
 
 .. currentmodule:: motor.motor_tornado
 
+Motor 2.3
+---------
+
+Motor 2.3 adds support for contextvars.
+
+New features:
+
+- Added supported for the contextvars module. Specifically, it is now possible
+  to access context variables inside
+  :class:`~pymongo.monitoring.CommandListener` callbacks.
+
+Bug-fixes:
+
+-
+
+Issues Resolved
+~~~~~~~~~~~~~~~
+
+See the `Motor 2.3 release notes in JIRA`_ for the complete list of resolved
+issues in this release.
+
+.. _Motor 2.3 release notes in JIRA: https://jira.mongodb.org/secure/ReleaseNote.jspa?projectId=11182&version=29749
+
 Motor 2.2
 ---------